% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ms_drive.R
\name{get_ms_drive}
\alias{get_ms_drive}
\title{Get the user's OneDrive for Business drive}
\usage{
get_ms_drive(tennant = NULL, app = NULL, pwd = NULL, email = NULL)
}
\arguments{
\item{tennant}{String indicating the Sharepoint tennant}

\item{app}{String Azure app id}

\item{pwd}{String Azure app secret token}

\item{email}{String MS email for the OneDrive user}
}
\description{
Authenticates with the Microsoft Graph API, and returns
an object representing the OneDrive of the given user.
}
