% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ms_table.R
\name{get_ms_table}
\alias{get_ms_table}
\title{Fetch experience from OneDrive database}
\usage{
get_ms_table(dir = NULL, file = NULL, table = NULL)
}
\arguments{
\item{dir}{R6 object of class ms_drive representing the target directory}

\item{file}{String filename of xlsx source}

\item{table}{String MS Graph ID for the Excel table within the xlsx file}
}
\description{
Performs a GET request to the Microsoft Graph API.
Returns a tibble containing the data within the specified
xlsx table.
}
